End to end
We are now able to execute queries in the cloak
and get results back that are persisted to disk.

+++ lib/proto/air/aggregate_results.pb.rb
@@ -23,7 +23,7 @@ class Percentile
 
     optional :percent, :uint32, 1
     optional :value, :sint64, 2
-    optional :percent_noise, :double, 3
+    optional :percent_noise, :float, 3
 
   end
